requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-environ==0.4.5
django-extensions==2.1.6
django-filter==2.1.0
django-ipware==2.1.0
djangorestframework==3.9.1
idna==2.8
Markdown==3.0.1
phpserialize==1.3
Pillow==5.4.1
pycrypto==2.6.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
